+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # Unreleased
 - Improve code diagnostics double-literal-format, no-boolean-literal-compare
 - Add experimental static code diagnostics newline-before-return, no-empty-block
+- Support cyclomatic complexity metric in analyzer plugin
 
 # 1.5.0
 - Add experimental static code diagnostics double-literal-format, no-boolean-literal-compare